[refactor] Repositoryを表すItemを別ファイルに分割
